+++ engine/source/scriptlib/ImportCache.c
@@ -121,7 +121,7 @@ HRESULT ImportNode_Init(ImportNode *self, const char *path)
 #ifdef IC_DEBUG
             fprintf(stderr, "ImportNode_Init: %s: %s@%i\n", path, List_GetName(list), i);
 #endif
-            List_InsertAfter(&self->functions, (void *)i, List_GetName(list));
+            List_InsertAfter(&self->functions, (void *)(size_t)i, List_GetName(list));
         }
         List_GotoNext(list);
     }
@@ -137,7 +137,7 @@ HRESULT ImportNode_Init(ImportNode *self, const char *path)
     size = List_GetSize(&self->functions);
     for(i = 0; i < size; i++)
     {
-        int index = (int)List_Retrieve(&self->functions);
+        int index = (size_t)List_Retrieve(&self->functions);
         List_Update(&self->functions, &(list->solidlist[index]));
         assert(((Instruction *)(list->solidlist[index]))->OpCode == FUNCDECL);
         List_GotoNext(&self->functions);
